feat: add shared JSON field handler
