+++ packages/shared-utils/src/index.ts
@@ -4,3 +4,4 @@ export { genSecret } from './genSecret';
 export { toggleItem } from './toggleItem';
 export { getCsrfToken } from './getCsrfToken';
 export { parseJsonBody } from './parseJsonBody';
+export { jsonFieldHandler, type ErrorSetter } from './jsonFieldHandler';
